// types.ts
export type SchemaFieldType = 'STRING' | 'NUMBER' | 'BOOLEAN' | 'ARRAY_OF_STRINGS' | 'OBJECT' | 'ARRAY_OF_OBJECTS';

export interface SchemaField {
  id: string;
  name: string;
  type: SchemaFieldType;
  children?: SchemaField[];
  error?: string;
}

export interface ExtractionResult {
    id: string;
    type: 'extraction' | 'transcription';
    fileId: string;
    fileName: string;
    timestamp: string;
    extractedData?: object; // Para extracciones
    schema?: SchemaField[];   // Para extracciones
    transcription?: string; // Para transcripciones
    metadata?: Metadata; // Para metadatos generados
}

export interface Metadata {
  title: string;
  summary: string;
  keywords: string[];
}